+++ package.json
@@ -42,10 +42,10 @@
     "react": "*"
   },
   "devDependencies": {
+    "@types/jest": "^25.2.1",
     "@types/lodash": "^4.14.135",
     "@types/react": "^16.8.19",
     "@types/react-dom": "^16.8.4",
-    "@types/warning": "^3.0.0",
     "enzyme": "^3.1.0",
     "enzyme-adapter-react-16": "^1.0.2",
     "enzyme-to-json": "^3.1.4",
@@ -63,7 +63,6 @@
   "dependencies": {
     "@babel/runtime": "^7.8.4",
     "async-validator": "^3.0.3",
-    "rc-util": "^4.17.0",
-    "warning": "^4.0.3"
+    "rc-util": "^4.20.3"
   }
 }
