+++ src/utils/validateUtil.ts
@@ -1,6 +1,6 @@
 import RawAsyncValidator from 'async-validator';
 import * as React from 'react';
-import warning from 'warning';
+import warning from 'rc-util/lib/warning';
 import {
   InternalNamePath,
   ValidateOptions,
