+++ tests/index.test.js
@@ -180,125 +180,6 @@ describe('Form.Basic', () => {
     });
   });
 
-  describe('initialValues', () => {
-    it('works', () => {
-      let form;
-
-      const wrapper = mount(
-        <div>
-          <Form
-            ref={instance => {
-              form = instance;
-            }}
-            initialValues={{ username: 'Light', path1: { path2: 'Bamboo' } }}
-          >
-            <Field name="username">
-              <Input />
-            </Field>
-            <Field name={['path1', 'path2']}>
-              <Input />
-            </Field>
-          </Form>
-        </div>,
-      );
-
-      expect(form.getFieldsValue()).toEqual({
-        username: 'Light',
-        path1: {
-          path2: 'Bamboo',
-        },
-      });
-      expect(form.getFieldsValue(['username'])).toEqual({
-        username: 'Light',
-      });
-      expect(form.getFieldsValue(['path1'])).toEqual({
-        path1: {
-          path2: 'Bamboo',
-        },
-      });
-      expect(form.getFieldsValue(['username', ['path1', 'path2']])).toEqual({
-        username: 'Light',
-        path1: {
-          path2: 'Bamboo',
-        },
-      });
-      expect(
-        getField(wrapper, 'username')
-          .find('input')
-          .props().value,
-      ).toEqual('Light');
-      expect(
-        getField(wrapper, ['path1', 'path2'])
-          .find('input')
-          .props().value,
-      ).toEqual('Bamboo');
-    });
-
-    it('update and reset should use new initialValues', () => {
-      let form;
-      let mountCount = 0;
-
-      const TestInput = props => {
-        React.useEffect(() => {
-          mountCount += 1;
-        }, []);
-
-        return <Input {...props} />;
-      };
-
-      const Test = ({ initialValues }) => (
-        <Form
-          ref={instance => {
-            form = instance;
-          }}
-          initialValues={initialValues}
-        >
-          <Field name="username">
-            <Input />
-          </Field>
-          <Field name="email">
-            <TestInput />
-          </Field>
-        </Form>
-      );
-
-      const wrapper = mount(<Test initialValues={{ username: 'Bamboo' }} />);
-      expect(form.getFieldsValue()).toEqual({
-        username: 'Bamboo',
-      });
-      expect(
-        getField(wrapper, 'username')
-          .find('input')
-          .props().value,
-      ).toEqual('Bamboo');
-
-      // Should not change it
-      wrapper.setProps({ initialValues: { username: 'Light' } });
-      wrapper.update();
-      expect(form.getFieldsValue()).toEqual({
-        username: 'Bamboo',
-      });
-      expect(
-        getField(wrapper, 'username')
-          .find('input')
-          .props().value,
-      ).toEqual('Bamboo');
-
-      // Should change it
-      form.resetFields();
-      wrapper.update();
-      expect(mountCount).toEqual(1);
-      expect(form.getFieldsValue()).toEqual({
-        username: 'Light',
-      });
-      expect(
-        getField(wrapper, 'username')
-          .find('input')
-          .props().value,
-      ).toEqual('Light');
-    });
-  });
-
   it('should throw if no Form in use', () => {
     const errorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
